# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Sport Booking Service API
  version: 0.1.0
  description: >
    API contract for the Ho Chi Minh City badminton-court marketplace MVP.
    Payments, memberships, reviews, and dynamic pricing are outside this version.
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /venues:
    get:
      tags: [Discovery]
      summary: Discover active venues
      operationId: discoverVenues
      security: []
      parameters:
        - $ref: "#/components/parameters/Latitude"
        - $ref: "#/components/parameters/Longitude"
        - name: radiusMeters
          in: query
          schema:
            type: integer
            minimum: 100
            maximum: 50000
            default: 5000
        - name: availableFrom
          in: query
          description: Include venues with at least one available court from this instant.
          schema:
            type: string
            format: date-time
        - name: availableTo
          in: query
          schema:
            type: string
            format: date-time
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Matching venues
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenuePage"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      tags: [Venues]
      summary: Create a venue
      operationId: createVenue
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVenueRequest"
      responses:
        "201":
          description: Venue created
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /venues/{venueId}:
    parameters:
      - $ref: "#/components/parameters/VenueId"
    get:
      tags: [Discovery]
      summary: Get venue details
      operationId: getVenue
      security: []
      responses:
        "200":
          description: Venue details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Venues]
      summary: Update an owned venue
      operationId: updateVenue
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVenueRequest"
      responses:
        "200":
          description: Venue updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /venues/{venueId}/courts:
    parameters:
      - $ref: "#/components/parameters/VenueId"
    get:
      tags: [Courts]
      summary: List courts at a venue
      operationId: listCourts
      security: []
      responses:
        "200":
          description: Courts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Court"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Courts]
      summary: Create a court at an owned venue
      operationId: createCourt
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourtRequest"
      responses:
        "201":
          description: Court created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Court"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /venues/{venueId}/operating-hours:
    parameters:
      - $ref: "#/components/parameters/VenueId"
    put:
      tags: [Venues]
      summary: Replace an owned venue's complete weekly operating schedule
      operationId: replaceVenueOperatingHours
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplaceOperatingHoursRequest"
      responses:
        "200":
          description: Complete replacement weekly schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueOperatingHours"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /courts/{courtId}:
    parameters:
      - $ref: "#/components/parameters/CourtId"
    patch:
      tags: [Courts]
      summary: Update a court at an owned venue
      operationId: updateCourt
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCourtRequest"
      responses:
        "200":
          description: Court updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Court"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /courts/{courtId}/blocks:
    parameters:
      - $ref: "#/components/parameters/CourtId"
    post:
      tags: [Courts]
      summary: Block a court interval
      operationId: createCourtBlock
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourtBlockRequest"
      responses:
        "201":
          description: Court interval blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourtBlock"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    VenueId:
      name: venueId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CourtId:
      name: courtId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Latitude:
      name: latitude
      in: query
      schema:
        type: number
        format: double
        minimum: -90
        maximum: 90
    Longitude:
      name: longitude
      in: query
      schema:
        type: number
        format: double
        minimum: -180
        maximum: 180
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Authentication is required or invalid
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: The caller cannot perform this operation
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: The request conflicts with current resource state
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    VenueStatus:
      type: string
      enum: [DRAFT, ACTIVE, INACTIVE]
    CourtStatus:
      type: string
      enum: [ACTIVE, INACTIVE, MAINTENANCE]
...
    Coordinates:
      type: object
      required: [latitude, longitude]
      properties:
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
    CreateVenueRequest:
      type: object
      required: [name, addressLine, district, location]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 160
        description:
          type: string
          maxLength: 2000
        phoneNumber:
          type: string
          maxLength: 30
        addressLine:
          type: string
          maxLength: 255
        ward:
          type: string
          maxLength: 120
        district:
          type: string
          maxLength: 120
        location:
          $ref: "#/components/schemas/Coordinates"
    UpdateVenueRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 160
        description:
          type: string
          maxLength: 2000
        phoneNumber:
          type: string
          maxLength: 30
        status:
          $ref: "#/components/schemas/VenueStatus"
    Venue:
      type: object
      required: [id, name, addressLine, district, city, location, timezone, status]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        phoneNumber:
          type: string
        addressLine:
          type: string
        ward:
          type: string
        district:
          type: string
        city:
          type: string
        location:
          $ref: "#/components/schemas/Coordinates"
        timezone:
          type: string
          example: Asia/Ho_Chi_Minh
        status:
          $ref: "#/components/schemas/VenueStatus"
        distanceMeters:
          type: integer
          minimum: 0
    VenueDetail:
      allOf:
        - $ref: "#/components/schemas/Venue"
        - type: object
          required: [courts]
          properties:
            courts:
              type: array
              items:
                $ref: "#/components/schemas/Court"
    VenuePage:
      type: object
      required: [items, page, size, totalItems, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Venue"
        page:
          type: integer
        size:
          type: integer
        totalItems:
          type: integer
          format: int64
        totalPages:
          type: integer
    CreateCourtRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 1000
    UpdateCourtRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 1000
        status:
          $ref: "#/components/schemas/CourtStatus"
    Court:
      type: object
      required: [id, venueId, name, status]
      properties:
        id:
          type: string
          format: uuid
        venueId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/CourtStatus"
    OperatingPeriod:
      type: object
      required: [dayOfWeek, opensAt, closesAt]
      properties:
        dayOfWeek:
          type: integer
          minimum: 1
          maximum: 7
        opensAt:
          type: string
          format: time
        closesAt:
          type: string
          format: time
    ReplaceOperatingHoursRequest:
      type: object
      required: [periods]
      properties:
        periods:
          type: array
          items:
            $ref: "#/components/schemas/OperatingPeriod"
    VenueOperatingHours:
      type: object
      required: [venueId, periods]
      properties:
        venueId:
          type: string
          format: uuid
        periods:
          type: array
          items:
            $ref: "#/components/schemas/OperatingPeriod"
    TimeInterval:
      type: object
      required: [startsAt, endsAt]
      properties:
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
...
    CreateCourtBlockRequest:
      allOf:
        - $ref: "#/components/schemas/TimeInterval"
        - type: object
          properties:
            reason:
              type: string
              maxLength: 255
    CourtBlock:
      allOf:
        - $ref: "#/components/schemas/CreateCourtBlockRequest"
        - type: object
          required: [id, courtId, createdByUserId, createdAt]
          properties:
            id:
              type: string
              format: uuid
            courtId:
              type: string
              format: uuid
            createdByUserId:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
...
    FieldViolation:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
    Problem:
      type: object
      required: [type, title, status, detail, instance]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri
        code:
          type: string
          description: Stable application error code.
        violations:
          type: array
          items:
            $ref: "#/components/schemas/FieldViolation"